back-end/requirements.txt
fastapi==0.73.0
pydantic==1.9.0
uvicorn==0.15.0
motor==2.4.0
pyjwt==2.3.0
python-multipart==0.0.5
boto3==1.17.106
botocore==1.20.106
python-dotenv==0.19.2
pymongo==3.12.3